producer/requirements.txt
asyncio==3.4.3
certifi==2019.11.28
kafka-python==1.4.6
numpy==1.18.1
pandas==1.0.1
protobuf==3.9.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.1
six==1.12.0
